% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AICc_manual}
\alias{AICc_manual}
\title{manually calculate AICc}
\usage{
AICc_manual(aic, n, k)
}
\arguments{
\item{aic}{Akaike Information Criterion calculated from AIC() or other functions, can be a vector}

\item{n}{the number of data points, can be a vector}

\item{k}{the number of parameters, can be a vector}
}
\value{
the AICc values in a numeric vector
}
\description{
calculate AICc from summary statistics
}
