% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_CI_from_Hessian_minlog}
\alias{calculate_CI_from_Hessian_minlog}
\alias{calculate_CI_from_variance_covariance}
\title{calculate CI from Hessian matrix}
\usage{
calculate_CI_from_Hessian_minlog(params, hessian, alpha = 0.05)

calculate_CI_from_variance_covariance(params, vcov, alpha = 0.05)
}
\arguments{
\item{params}{fitted parameters returned by optimizer}

\item{hessian}{the hessian matrix returned by optimizer}

\item{alpha}{desired confidence level}

\item{vcov}{the variance-covariance matrix returned by optimizer}
}
\value{
\code{calculate_CI_from_Hessian_minlog} and \code{calculate_CI_from_variance_covariance} return a data frame listing estimated parameters and their upper and lower CI
}
\description{
Calculate CI from Hessian matrix of a minus-log-likelihood function
}
