% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powbend.R
\name{powbend_term}
\alias{powbend_term}
\alias{d_powbend}
\alias{p_powbend}
\alias{q_powbend}
\alias{r_powbend}
\alias{fit_powbend}
\title{Probability functions of power-bend distribution}
\usage{
powbend_term(x, s, lambda, adjust = 0, log = FALSE)

d_powbend(
  x,
  s,
  lambda = 0.001,
  minlogLambda = -log(lambda),
  trunc = NULL,
  exact = 1e+05,
  log = FALSE
)

p_powbend(
  q,
  s,
  lambda = 0.001,
  minlogLambda = -log(lambda),
  trunc = NULL,
  exact = 1e+05,
  log.p = FALSE,
  lower.tail = TRUE
)

q_powbend(
  p,
  s,
  lambda = 0.001,
  minlogLambda = -log(lambda),
  trunc = NULL,
  exact = 1e+05,
  log.p = FALSE,
  lower.tail = TRUE
)

r_powbend(
  n,
  s,
  lambda = 0.001,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  trunc = NULL
)

fit_powbend(x, trunc = NULL, exact = 1e+05, start.value, ...)
}
\arguments{
\item{x, q}{vector of quantiles (abundances or read numbers in a sample)}

\item{s}{exponent of the denominator in the power-bend distribution}

\item{lambda, minlogLambda}{the bending parameter of the distribution (scaling coefficient of the exponent in numerator)}

\item{adjust}{normalizing coefficient for integration of very large or small values}

\item{log, log.p}{logical;if the probabilities returned by or supplied to the function are in log-scale}

\item{trunc}{truncation point of the distribution; defaults to \code{NULL}}

\item{exact}{the number of exact sum to be calculated in the infinite series}

\item{lower.tail}{logical;if the cumulative probabilities are lower-tail (\code{TRUE}) or upper-tail (\code{FALSE})}

\item{p}{vector of cumulative probabilities}

\item{n}{the number of random values to be generated}

\item{start.value}{the starting values of fitting a power-bend distribution}

\item{...}{other parameters supplied to the \code{mle2} function}
}
\value{
\code{powbend_term},\code{d_powbend},\code{p_powbend},\code{r_powbend},\code{q_powbend} returns a numeric vector

\code{fit_powbend} returns a object of \code{mle2} class
}
\description{
Alternative implementation of power-bend distribution functions with faster speed and better accuracy for large x or q
}
\details{
The \verb{[dpqr]_powbend} functions are counterparts of the \verb{[dpqr]powbend} functions in the \code{sads} package with better support and faster computation time for s<1 and large quantiles.

The \code{powbend_terms} function calculates the unnormalized terms in a power-bend distribution that are useful for many distribution functions based on power-bend distribution.

The \code{fit_powbend} function a counterpart of \code{sads::fitpowbend} that uses \code{d_powbend} function.
}
