% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbpowbend.R
\name{fit_nbpowbend}
\alias{fit_nbpowbend}
\title{Fit SAD with negative-binomial-power-bend distribution}
\usage{
fit_nbpowbend(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  exact = 1e+05,
  constant = "G25_K51",
  start.value,
  min.odisp = 0.001,
  min.eta = 1e-16,
  ...
)
}
\arguments{
\item{x}{the observed species abundances}

\item{gcn}{the 16S copy number of the}

\item{tol}{the maximum relative error of logarithm;if exceeded, a warning message will be returned}
}
\value{
returns a boolean vector
}
\description{
Fit a power-bend SAD compounded with negative-binomial distribution
}
\details{
This function is cloned frminlogLambda the internal function frminlogLambda \code{sads} package
}
