% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incomplete_gamma.R
\name{get_incgam}
\alias{get_incgam}
\title{Calculate incomplete gamma function}
\usage{
get_incgam(a, x, log = FALSE, debug = FALSE)
}
\arguments{
\item{a}{exponent of the power-law term}

\item{x}{the starting point of the infinite integration}

\item{log}{if \code{TRUE}, log-transformed value is returned; defaults to \code{FALSE}}
}
\value{
\code{get_incgam} returns a single numeric value
}
\description{
Calculate upper incomplete gamma function using recursive equation
}
\details{
\code{get_incgam} is an extension from \code{pracma::incgam} to handle situations where \code{a} is very large or negative

It uses the recurrence relation: incgam(x,a+1) = incgam(x,a)a + exp(-x)x^a

It is NOT vectorized for either \code{a} or \code{x}
}
