% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Integer check}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the numbers to be checked}

\item{tol}{the tolerance to distinguish real number from an integer; defaults to the square-root of the double-float precision as in the \code{sads} package.}
}
\value{
\code{is.wholenumber} returns a Boolean vector
}
\description{
Integer check for discrete distributions and other discrete functions
}
\details{
The function \code{is.wholenumber} is cloned from the internal function in \code{sads} package. It is exported so that user's customized SAD functions may use it.

This function is vectorized.
}
