% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral_refined_02inf_lnorm}
\alias{integral_refined_02inf_lnorm}
\title{Specialized infinite integration of lognormal based SAD by expanding intervals}
\usage{
integral_refined_02inf_lnorm(
  fun,
  params,
  fin.intervals = qnorm(seq(5e-04, 0.9995, length.out = 9)),
  constant = "G25_K51",
  no_intervals = 8,
  random = FALSE,
  tol = 1e-08,
  silent = FALSE,
  cutoff = -19,
  nCheck = 100,
  log = FALSE
)
}
\arguments{
\item{fun}{the function for the integrand}

\item{params}{the parameters for the integrand}

\item{fin.intervals}{the logged breaking points for the finite intervals}

\item{constant}{the constants used in Gauss-Kronrod quadrature}

\item{no_intervals}{the number of subdivisions at start; the same as in \code{pracma::integral}}

\item{random}{logical; shall the length of subdivisions be random; the same as in \code{pracma::integral}}

\item{tol}{relative tolerance}

\item{silent}{logical; whether to output warnings and informative messages}

\item{cutoff}{the logged threshold of integration termination; not used for lognormal-based integrands}

\item{nCheck}{the number of subdivisions of a finite interval for normalization of integrand values}

\item{log}{logical; whether to return the logged value of the integral}
}
\value{
\code{integral_refined_a2inf_lnorm} returns a single numeric value for the integration.
}
\description{
A specialized infinite integral function from 0 to Inf for use in lognormal based SAD probability functions
}
\details{
\code{integral_refined_a2inf_lnorm} is a specialized protocol of integration for distribution functions like \code{dpoilog} for desired numerical accuracy in a reasonable time.

It is designed to minimize numerical instability associated with \code{pracma::quadinf} by dividing the interval to smaller intervals.
}
