% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral_refined_a2inf}
\alias{integral_refined_a2inf}
\title{Specialized infinite integration by expanding intervals}
\usage{
integral_refined_a2inf(
  fun,
  params,
  xmin,
  xbreak,
  fold = 10,
  constant = "G25_K51",
  no_intervals = 8,
  random = FALSE,
  tol = 1e-08,
  silent = FALSE,
  cutoff = -19,
  cutoff.type = c(1, 2, 3),
  nCheck = 100,
  log = FALSE
)
}
\arguments{
\item{fun}{the function for the integrand}

\item{params}{the parameters for the integrand}

\item{xmin}{the starting point of integration}

\item{xbreak}{the end point of the first segmented integration}

\item{fold}{the coefficient of exponential expansion of the interval of integration}

\item{constant}{the constants used in Gauss-Kronrod quadrature}

\item{no_intervals}{the number of subdivisions at start; the same as in \code{pracma::integral}}

\item{random}{logical; shall the length of subdivisions be random; the same as in \code{pracma::integral}}

\item{tol}{relative tolerance}

\item{silent}{logical; whether to output warnings and informative messages}

\item{cutoff}{the logged threshold of integration termination}

\item{cutoff.type}{the metrics used for intergration termination, see \code{Details}}

\item{nCheck}{the number of subdivisions of a finite interval for normalization of integrand values}

\item{log}{logical; whether to return the logged value of the integral}
}
\value{
\code{integral_refined_a2inf} returns a single numeric value for the integration.
}
\description{
A specialized infinite integral function from a to Inf for use in probability functions
}
\details{
\code{integral_refined_a2inf} is a specialized protocol of integration for distribution functions like \code{dnbpowbend} for desired numerical accuracy in a reasonable time.

It is designed to minimize numerical instability associated with \code{pracma::quadinf} for some integrands by dividing the infinite interval to exponentially expanding finite intervals.

Three metrics are used to determine whether such expansion may stop:

\enumerate{
\item if the integral value (finite+infinite) remains stable across iterations
}

\enumerate{
\item if the infinite integral value becomes negligible to the sum of finite integrals
}

\enumerate{
\item if the increment of finite integral value becomes negligible
}

Not all 3 metrics are useful (and sometimes one may be harmful) for certain integrands, and user may choose which metric to use by supplying a numeric vector of the metric indices through \code{cutoff.type}.
}
