% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral_refined_a2inf_power}
\alias{integral_refined_a2inf_power}
\title{Specialized infinite integration of power-law based SAD by expanding intervals}
\usage{
integral_refined_a2inf_power(
  fun,
  q,
  s,
  eta,
  odisp,
  xmin,
  xbreak,
  fold = 10,
  constant = "G25_K51",
  no_intervals = 8,
  random = FALSE,
  tol = 1e-08,
  silent = FALSE,
  cutoff = -19,
  cutoff.type = c(1, 2),
  nCheck = 100,
  log = FALSE
)
}
\arguments{
\item{fun}{the function for the integrand}

\item{q, s, eta, odisp}{the parameters of the power-law-based integrand}

\item{xmin}{the starting point of integration}

\item{xbreak}{the end point of the first segmented integration}

\item{fold}{the coefficient of exponential expansion of the interval of integration}

\item{constant}{the constants used in Gauss-Kronrod quadrature}

\item{no_intervals}{the number of subdivisions at start; the same as in \code{pracma::integral}}

\item{random}{logical; shall the length of subdivisions be random; the same as in \code{pracma::integral}}

\item{tol}{relative tolerance}

\item{silent}{logical; whether to output warnings and informative messages}

\item{cutoff}{the logged threshold of integration termination}

\item{cutoff.type}{the metrics used for intergration termination, see \code{Details}}

\item{nCheck}{the number of subdivisions of a finite interval for normalization of integrand values}

\item{log}{logical; whether to return the logged value of the integral}
}
\value{
\code{integral_refined_a2inf_power} returns a single numeric value for the integration.
}
\description{
A specialized infinite integral function from a to Inf for use in power-law based SAD probability functions
}
\details{
\code{integral_refined_a2inf_power} is a specialized protocol of integration for distribution functions like \code{poipower} for desired numerical accuracy in a reasonable time.

It is designed to minimize numerical instability associated with \code{pracma::quadinf} by dividing the infinite interval to exponentially expanding finite intervals.

Two metrics are used to determine whether such expansion may stop:

\enumerate{
\item if the integral value (finite+infinite) remains stable across iterations
}

\enumerate{
\item if the finite integral values comply to a geometric series so analytic solutions of their sum may be used
}

User may choose which metric to use by supplying a numeric vector of the metric indices through \code{cutoff.type}.
}
