% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral_refined}
\alias{integral_refined}
\title{Integration with Gauss-Kronrod Quadrature}
\usage{
integral_refined(
  fun,
  xmin,
  xmax,
  constant = "G25_K51",
  no_intervals = 8,
  random = FALSE,
  tol = 1e-08,
  silent = FALSE,
  ...
)
}
\arguments{
\item{fun}{integrand, univariate (vectorized) function; the same as in \code{pramca::integral}}

\item{xmin, xmax}{endpoints of the integration interval; the same as in \code{pramca::integral}}

\item{constant}{a character name for the Gauss-Kronrod system or a list of user-supplied numeric vectors; see \code{Details} for more information}

\item{no_intervals}{the number of subdivisions at start; the same as in \code{pracma::integral}}

\item{random}{logical; shall the length of subdivisions be random; the same as in \code{pracma::integral}}

\item{tol}{relative tolerance}

\item{silent}{logical; whether to output warnings and informative messages}

\item{...}{additional parameters to be passed to the integrand function; the same as in \code{pracma::integral}}
}
\value{
\code{integral_refined} returns a single numeric value for the integration.
}
\description{
A wrapper function for integration through Gauss-Kronrod (quadgk) or tanh-sinh quadrature (quadinf)
}
\details{
\code{integral_refined} is a modified version of \code{pracma::integral} to allow the use of higher-order GK nodes for finite integration by the \code{constant} parameter.

The \code{constant} parameter takes either a character name from names(quadgk_constant) or a list of user-provided node constants.

If provided by the user, \code{constant} should be a list of four elements with the name \code{K_node},\code{K_weight},\code{G_node},\code{G_weight}.
}
\seealso{
\code{\link[=quadgk_refined]{quadgk_refined()}}
}
