% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lerch_transcendent.R
\name{Lerch_transcendent_approximation}
\alias{Lerch_transcendent_approximation}
\alias{Lerch_transcendent_tail_approx}
\title{Calculate Lerch transcendent for log-series and power-bend distribution}
\usage{
Lerch_transcendent_approximation(method = "_tail_approx", ...)

Lerch_transcendent_tail_approx(
  x = 1,
  lambda,
  s,
  exact = 1e+05,
  log = FALSE,
  debug = FALSE
)
}
\arguments{
\item{x}{the index of element in the log-series or power-bend after which the sum is calculated}

\item{lambda}{the base of the numerator; \code{exp(-lambda)} is identical to \code{z} in \code{pracma::polylog}}

\item{s}{the exponent of the denominator; \code{s} is identical to \code{n} in \code{pracma::polylog}}

\item{exact}{the number of terms that is added exactly before approximation applies}

\item{log}{if \code{TRUE}, return the log-transformed value; defaults to \code{FALSE}}
}
\value{
\code{Lerch_transcendent_tail_approx} returns a single numeric value
}
\description{
Calculate Lerch transcendent for log-series and power-bend with better support for base close to 1
}
\details{
The Lerch transcendent function is particularly useful for calculating power-bend CDF and the approximated q-function

\code{Lerch_transcendent_approximation} is the wrapper function for Lerch transcendent function

\code{Lerch_transcendent_tail_approx} approximates the Lerch transcendent. The tail part of the infinite sum of the series is approximated by the continuous incomplete gamma function. Half of the \code{exact+1}th element in the series is supplemented to compensate the continuous approximation.
}
