% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_arithmetic.R
\name{log_x_plus_y}
\alias{log_x_plus_y}
\alias{log_sum_x}
\alias{log_one_minus_x}
\title{Arithmetic under log-scale}
\usage{
log_x_plus_y(x, y)

log_sum_x(x)

log_one_minus_x(x)
}
\arguments{
\item{x}{values to be added or subtracted}

\item{y}{values to be added}
}
\value{
\code{log_x_plus_y} and \code{log_sum_x} return a single numeric value; \code{log_one_minus_x} returns a numeric vector with the same length as input \code{x}
}
\description{
Add or subtract extremely large or small numbers in log-scale to avoid numerical issues
}
\details{
\code{log_x_plus_y} calculates z so that exp(z) = exp(x)+exp(y)

\code{log_sum_x} calculates z so that exp(z) = sum(exp(x))

\code{log_one_minus_x} calculates z so that exp(z) = 1-exp(x)

Only \code{log_one_minus_x} is vectorized
}
