% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbpowbend.R
\name{dtrunc_dnbpowbend}
\alias{dtrunc_dnbpowbend}
\alias{dnbpowbend}
\alias{dconv_nbpowbend}
\alias{pnbpowbend}
\alias{pconv_nbpowbend}
\alias{rnbpowbend}
\title{Negative-binomial-power-bend distribution functions}
\usage{
dtrunc_dnbpowbend(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  eta = 0.001,
  odisp,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  constant = "G25_K51",
  log = FALSE,
  debug = FALSE
)

dnbpowbend(
  x,
  gcn = rep(1, length(x)),
  eta = 0.001,
  odisp,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  constant = "G25_K51",
  log = FALSE,
  debug = FALSE
)

dconv_nbpowbend(x, q, s, lambda, eta, odisp, adjust = 0, log = FALSE)

pnbpowbend(
  q,
  gcn = rep(1, length(q)),
  eta = 0.001,
  odisp,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  constant = "G25_K51",
  log.p = FALSE,
  lower.tail = FALSE,
  debug = FALSE
)

pconv_nbpowbend(x, q, s, lambda, eta, odisp, adjust = 0, log = FALSE)

rnbpowbend(
  n,
  gcn = rep(1, n),
  eta = 0.001,
  odisp,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda)
)
}
\arguments{
\item{x, q}{the observed species abundances; in 16S profiling, this is the observed 16S read count}

\item{gcn}{the 16S copy number of OTUs; it should be a numeric vector of the same length as \code{x}}

\item{trunc}{the truncation point to fit a distribution whose lower-tail is truncated; non-negative integer}

\item{eta}{the sampling effort measured as the expected number of reads per individual}

\item{odisp}{the over-dispersion parameter \code{r} or \code{size} in the negative-binomial distribution that represents sampling bias}

\item{s}{the s parameter of power-bend distribution; it is the same as parameter s in \code{sads} package}

\item{lambda, minglogLambda}{the lambda parameter of power-bend distribution; it is the same as parameter omega and oM in \code{sads} package}

\item{exact}{the number of terms to be calculated exactly before approximation applies}

\item{constant}{the name of Gauss-Kronrod quadrature formula; See details}

\item{log, log.p}{if log-transformed probability should be returned or received; default to FALSE}

\item{adjust}{log-adjustment value to avoid precision problem in integration}

\item{lower.tail}{if the lower-tail cumulative probabilty should be returned or received; default to FALSE for accuracy}

\item{n}{the number of random numbers to be generated}

\item{start.value}{a list of named numeric values as the starting values of ML fitting}

\item{...}{other parameters passed to the \code{optim} function}
}
\value{
\code{dnbpowbend},\code{pnbpowbend} and \code{dtrunc_dnbpowbend} returns a numeric vector of the (log) probability

\code{rnbpowbend} returns a numeric vector of the random numbers

\code{fit_nbpowbend} returns a \code{mle2} class object containing the fitted model
}
\description{
The PMF, CDF and RNG functions of power-bend distributed communities with negative-binomial sampling error.
}
\details{
\code{dnbpowbend} and \code{dtrunc_dnbpowbend} calculates the full and truncated PMF for negative-binomial-power-bend distribution

\code{pnbpowbend} calculates the CDF of negative-binomial-power-bend distribution. To achieve the best accuracy for SAD fitting, it calculates the upper tail of the CDF by default

\code{rnbpowbend} generates random number under the negative-binomial-power-bend distribution by generating random number under the power-bend distribution and then sampling under negative-binomial distribution

\code{constant} denotes the name of a set of values used by Gauss-Kronrod quadrature. This packages provides two sets of these values: \code{G25_K51}, the default; and \code{G7_K15}, the same as used in the GK method in \code{pracma::integral}

\code{dcon_nbpowbend} and \code{pconv_nbpowbend} are the integrand functions used internally by \code{dnbpowbend} and \code{pnbpowbend}

\code{fit_nbpowbend} fits the negative-binomial-power-bend distribution to observed SAD with maximum likelihood and a default truncation point of 0 using Nelder-Mead optimum searching algorithm. Correction for 16S rRNA gene copy number can be done by supplying the corresponding GCN to the \code{gcn} parameter.

The quantile function \code{qnbpowbend} is not implemented and thus \code{rnbpowbend} only generates random values under the full distribution, making it not suitable for RAD prediction.

To predict RAD, use \code{representative_nbpowbend} for approximated quantiles.
}
