% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poipowbend.R
\name{dtrunc_dpoipowbend}
\alias{dtrunc_dpoipowbend}
\alias{dpoipowbend}
\alias{ppoipowbend}
\alias{pconv_poipowbend}
\alias{rpoipowbend}
\alias{fit_poipowbend}
\title{Power-bend SAD with Poisson sampling error}
\usage{
dtrunc_dpoipowbend(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  eta = 0.001,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  constant = "G25_K51",
  log = FALSE,
  debug = FALSE
)

dpoipowbend(
  x,
  gcn = rep(1, length(x)),
  eta = 0.001,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  log = FALSE,
  debug = FALSE
)

ppoipowbend(
  q,
  gcn = rep(1, length(q)),
  eta = 0.001,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  exact = 1e+05,
  constant = "G25_K51",
  log.p = FALSE,
  lower.tail = TRUE,
  debug = FALSE
)

pconv_poipowbend(x, q, s, lambda, eta, adjust = 0, log = FALSE)

rpoipowbend(
  n,
  gcn = rep(1, n),
  eta = 0.001,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda)
)

fit_poipowbend(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  exact = 1e+05,
  start.value,
  ...
)
}
\arguments{
\item{x, q}{the observed species abundances, non-negative integers}

\item{gcn}{the 16S copy number of each species, with the same length and order of x, positive integers}

\item{trunc}{the truncation point to fit a distribution whose lower-tail is truncated; non-negative integer}

\item{eta}{the sampling effort, positive real number}

\item{s}{the s parameter of power-bend distribution; it is the same as parameter s in \code{sads} package}

\item{lambda, minglogLambda}{the lambda parameter of power-bend distribution; it is the same as parameter omega and oM in \code{sads} package}

\item{exact}{the number of terms to be calculated exactly before approximation applies}

\item{constant}{the name of Gauss-Kronrod quadrature formula; See details}

\item{log, log.p}{if log-transformed probability should be returned or received; default to FALSE}

\item{lower.tail}{if the lower-tail cumulative probabilty should be returned or received; default to FALSE for accuracy}

\item{adjust}{log-adjustment value to avoid precision problem in integration}

\item{n}{the number of random numbers to be generated}

\item{start.value}{a list of named numeric values as the starting values of ML fitting}

\item{...}{other parameters passed to the \code{optim} function}
}
\value{
\code{dpoipowbend},\code{ppoipowbend} and \code{dtrunc_dpoipowbend} returns a numeric vector of the (log) probability

\code{rpoipowbend} returns a numeric vector of the random numbers

\code{fit_poipowbend} returns a \code{mle2} class object containing the fitted model
}
\description{
The PMF, CDF, RNG and fitting function of power-bend distributed communities with Poisson sampling error.
}
\details{
\code{dpoipowbend} and \code{dtrunc_dpoipowbend} calculates the full and truncated PMF for Poisson-power-bend distribution

\code{ppoipowbend} calculates the CDF of Poisson-power-bend distribution. To achieve the best accuracy for SAD fitting, it calculates the upper tail of the CDF by default

\code{rpoipowbend} generates random number under the Poisson-power-bend distribution by generating random number under the power-bend distribution and then sampling under Poisson distribution

\code{constant} denotes the name of a set of values used by Gauss-Kronrod quadrature. This packages provides two sets of these values: \code{G25_K51}, the default; and \code{G7_K15}, the same as used in the GK method in \code{pracma::integral}

\code{pconv_poipowbend} is the integrand function used internally by \code{ppoipowbend}

\code{fit_poipowbend} fits the Poisson-power-bend distribution to observed SAD with maximum likelihood and a default truncation point of 0 using Nelder-Mead optimum searching algorithm. Correction for 16S rRNA gene copy number can be done by supplying the corresponding GCN to the \code{gcn} parameter.

The quantile function \code{qpoipowbend} is not implemented and thus \code{rpoipowbend} only generates random values under the full distribution, making it not suitable for RAD prediction.

To predict RAD, use \code{representative_poipowbend} for approximated quantiles.
}
