% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poipower.R
\name{dtrunc_dpoipower}
\alias{dtrunc_dpoipower}
\alias{dpoipower}
\alias{ppoipower}
\alias{pconv_poipower}
\alias{fit_poipower}
\alias{qpoipower}
\alias{rpoipower}
\title{Power-law SAD with Poisson sampling error}
\usage{
dtrunc_dpoipower(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  eta,
  s,
  exact = 1e+05,
  constant = "G25_K51",
  log = FALSE,
  debug = FALSE
)

dpoipower(
  x,
  gcn = rep(1, length(x)),
  eta,
  s,
  exact = 1e+05,
  log = FALSE,
  debug = FALSE
)

ppoipower(
  q,
  gcn = rep(1, length(q)),
  eta,
  s,
  exact = 1e+05,
  constant = "G25_K51",
  log.p = FALSE,
  lower.tail = TRUE,
  debug = FALSE
)

pconv_poipower(x, q, eta, s, log = FALSE, adjust = 0)

fit_poipower(
  x,
  gcn = rep(1, length(x)),
  trunc = 0,
  start.value,
  upper = c(s = 20),
  ...
)

qpoipower(
  p,
  s,
  eta,
  trunc = 0,
  log.p = FALSE,
  lower.tail = FALSE,
  exact = 1e+05,
  constant = "G25_K51",
  Nmax = .Machine$integer.max,
  debug = FALSE
)

rpoipower(
  n,
  s,
  eta,
  trunc = 0,
  exact = 1e+05,
  constant = "G25_K51",
  Nmax = .Machine$integer.max,
  debug = FALSE
)
}
\arguments{
\item{x, q}{the observed species abundances, non-negative integers}

\item{gcn}{the 16S copy number of each species, with the same length and order of x, positive integers}

\item{trunc}{the truncation point to fit a distribution whose lower-tail is truncated; non-negative integer}

\item{eta}{the sampling effort, positive real number}

\item{s}{exponent of the denominator in power-law distribution; it is the same as parameter \code{s} in the \code{sads} package}

\item{exact}{the number of terms to be calculated exactly before approximation applies}

\item{constant}{the name of Gauss-Kronrod quadrature constants; see \code{Details}}

\item{log, log.p}{if log-transformed probability should be returned or supplied; default to \code{FALSE}}

\item{lower.tail}{if the lower-tail cumulative probability should be returned or received; default to \code{FALSE} for accuracy of large quantiles}

\item{adjust}{log-adjustment value to avoid precision problem in integration}

\item{start.value}{a list of named numeric values as the starting values of ML fitting}

\item{upper}{the upper limit of parameters in the ML fitting}

\item{...}{other parameters passed to the \code{mle2} function}

\item{p}{the cumulative probabilities of the species abundances, real values from 0 to 1}

\item{Nmax}{the maximum abundance beyond which the RAD is approximated by quantiles in a pareto distribution}

\item{n}{the number of random values to be generated}
}
\value{
\code{dpoipower},\code{ppoipower} and \code{dtrunc_poipower} returns a numeric vector of the (log) probability

\code{qpoipower},\code{rpoipower} returns a numeric vector of the (random) quantiles

\code{fitpoipower} returns a \code{mle2} class object containing the fitted model
}
\description{
The PMF, CDF, RNG and fitting function of power-law distributed communities with Poisson sampling error.
}
\details{
\code{dpoipower} and \code{dtrunc_dpoipower} calculates the full and truncated PMF for Poisson-power-law distribution

\code{ppoipower} calculates the CDF of Poisson-powerlaw distribution. To achieve the best accuracy for SAD fitting, it caculates the upper tail of the CDF by default

\code{qpoipower} calculates the quantile function for the Poisson-power-law distribution. Given a probability value, it estimates the corresponding value of the distribution at that probability. Large quantiles are approximated using the continuous pareto distribution.

\code{rpoipower} generates random numbers under the Poisson-power-law distribution.

\code{constant} denotes the name of a set of values used by Gauss-Kronrod quadrature. This packages provides two sets of these values: \code{G25_K51}, the default; and \code{G7_K15}, the same as used in the GK method in \code{pracma::integral}

\code{pconv_poipower} is the integrand function used internally by \code{ppoipower}

\code{fit_poipower} fits the Poisson-power-law distribution to observed SAD using maximum likelihood. Correction for 16S rRNA gene copy number can be done by supplying the corresponding GCN to the \code{gcn} parameter.
}
