% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polylog.R
\name{polylog_approximation}
\alias{polylog_approximation}
\alias{polylog_tail_approx}
\title{Calculate polylogarithm for log-series and power-bend distribution}
\usage{
polylog_approximation(method = "_tail_approx", ...)

polylog_tail_approx(lambda, s, exact = 1e+05, debug = FALSE, log = FALSE)
}
\arguments{
\item{lambda}{the base of the numerator; \code{exp(-lambda)} is identical to \code{z} in \code{pracma::polylog}}

\item{s}{the exponent of the denominator; \code{s} is identical to \code{n} in \code{pracma::polylog}}

\item{exact}{the number of terms that is added exactly before approximation applies}

\item{log}{if \code{TRUE}, return the log-transformed value; defaults to \code{FALSE}}
}
\value{
\code{polylog_tail_approx} returns a single numeric value
}
\description{
Calculate polylogarithm for log-series and power-bend with better support for base close to 1 and relevant arguments
}
\details{
\code{polylog_approximation} is the wrapper function for polylogarithm function

\code{polylog_tail_approx} calculates the polylogarithm with better support for \code{z=exp(-lambda)} close to 1 compared to pracma::polylog. The tail part of the infinite sum is approximated by the continuous incomplete gamma function. Half of the \code{exact+1}th element in the series is supplemented to compensate the continuous approximation.
}
