% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microMETER_data.R, R/microSAD_data.R
\docType{data}
\name{quadgk_constant}
\alias{quadgk_constant}
\title{Node and weight values for Gauss-Kronrod quadrature formula in quadgk}
\format{
A named list with 2 elements, with each element is a named list of 4 numeric vectors

\describe{
\item{G7_K15}{The node and weight values for a G7_K15 Gauss-Kronrod quadrature formula}
\item{G25_K51}{The node and weight values for a G25_K51 Gauss-Kronrod quadrature formula}
\item{$K_node}{The node values for Kronrod rule}
\item{$K_weight}{The weight values for Kronrod rule}
\item{$G_node}{The node values for Gauss rule}
\item{$G_weight}{The weight values for Gauss rule}
}

A named list with 2 elements, with each element is a named list of 4 numeric vectors

\describe{
\item{G7_K15}{The node and weight values for a G7_K15 Gauss-Kronrod quadrature formula}
\item{G25_K51}{The node and weight values for a G25_K51 Gauss-Kronrod quadrature formula}
\item{$K_node}{The node values for Kronrod rule}
\item{$K_weight}{The weight values for Kronrod rule}
\item{$G_node}{The node values for Gauss rule}
\item{$G_weight}{The weight values for Gauss rule}
}
}
\source{
All values are cited from Pavel Holoborodko, 2011
\url{https://www.advanpix.com/2011/11/07/gauss-kronrod-quadrature-nodes-weights}

All values are cited from Pavel Holoborodko, 2011
\url{https://www.advanpix.com/2011/11/07/gauss-kronrod-quadrature-nodes-weights}
}
\usage{
quadgk_constant

quadgk_constant
}
\description{
A named list of 2 elements, with each element is a named list of 4 numeric vectors

A named list of 2 elements, with each element is a named list of 4 numeric vectors
}
\keyword{datasets}
