% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{quadgk_refined}
\alias{quadgk_refined}
\title{Adaptive Gauss-Kronrod Quadrature with user-determined nodes and weights}
\usage{
quadgk_refined(
  f,
  a,
  b,
  constant = microSAD::quadgk_constant$G25_K51,
  tol = .Machine$double.eps^0.5,
  ...
)
}
\arguments{
\item{f}{the same as in \code{pracma::quadgk}: integrand as function; needs to be vectorized.}

\item{a, b}{the same as in \code{pracma::quadgk}: endpoints of the integration interval.}

\item{constant}{a list of node and weight values; see \code{Details} for more information.}

\item{tol}{the same as in \code{pracma::quadgk}: relative tolerance.}

\item{...}{the same as in \code{pracma::quadgk}: additional parameters to be passed to the function \code{f}.}
}
\value{
\code{quadgk_refined} returns a single numeric value for the integration.
}
\description{
A modified version of \code{pracma::quadgk} to use user-determined nodes and weights for Gauss-Kronrod quadrature
}
\details{
\code{quadgk_refined} is extended from \code{pracma::quadgk} function to use user-determined G(n)_K(2n+1) Gauss-Kronrod quadrature.

\code{constant} has to be a named list of 4 elements: \verb{$K_node},\verb{$K_weight},\verb{$G_node},\verb{$G_weight}

\verb{$K_node} and \code{K_weight} are numeric vectors of the nodes and weights for the Kronrod rule, with a length of 2n+1

\verb{$G_node} and \code{G_weight} are numeric vectors of the nodes and weights for the Gauss rule, with a length of n

\code{microMETER} package includes two sets of node and weight values in exported data \code{quadgk_constant}, in which \verb{$G25_K51} is used as the default set for \code{constant}.
}
\seealso{
\code{\link[=integral_refined]{integral_refined()}}
}
