% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lack_of_fit_test.R
\name{RAD_lack_of_fit_test}
\alias{RAD_lack_of_fit_test}
\title{Test the significance of lack of fit in SAD/RAD}
\usage{
RAD_lack_of_fit_test(
  obs,
  model,
  params,
  log = TRUE,
  n = 100,
  details = FALSE,
  debug = FALSE
)
}
\arguments{
\item{obs}{the observed abundances as in RAD; no need to be sorted}

\item{model}{the name of the model as appeared in the name of fit_* function}

\item{params}{the fitted parameters as returned by fit_* function plus additional arguments like \code{trunc}}

\item{log}{logical; whether the goodness/lack of fit is calculated at log-scale; defaults to \code{TRUE}}

\item{n}{the number of Monte-Carlo experiments to run for estimating the p-value}

\item{details}{logical; whether the randomly sampled RADs from Monte-Carlo experiments are returned; defaults to \code{FALSE}}
}
\value{
\code{RAD_lack_of_fit_test} returns a list of 3 or 4 elements containing the test statistics and other information

\verb{$R2} the observed R^2

\verb{$p.value} the p-value of obtaining the observed R^2

\verb{$representative} the representative RAD given the fitted SAD model

\verb{$random} optional; the random RADs sampled in the Monte-Carlo experiments
}
\description{
Test the significance of lack of fit in SAD/RAD through Monte-Carlo experiments
}
\details{
Although model selection using AIC can find the best model from a set of SAD models, it is not guaranteed that the best model found is sufficiently good to described the observed SAD/RAD.

Therefore, \code{RAD_lack_of_fit_test} compares the fitted SAD with the observed one by calculating the modified R^2 between the observed RAD and the representative RAD given by the fitted SAD.

The significance of lack of fit is estimated through Monte-Carlo experiments where random RADs are sampled from the fitted SAD and compared to the representative RAD to calculate the baseline empirical distribution for R^2.

Then the p-value is estimated as the empirical probability of obtaining the observed R^2 or a worse one under the baseline R^2 distribution.
}
