% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/representative_abundance.R
\name{representative_power}
\alias{representative_power}
\alias{representative_poipower}
\alias{representative_ls}
\alias{representative_powbend}
\alias{representative_poilog}
\alias{representative_nblog}
\alias{representative_poipowbend}
\alias{representative_nbpowbend}
\alias{representative_poils}
\alias{representative_nbls}
\title{Generate RAD from SAD model}
\usage{
representative_power(
  n,
  trunc = NULL,
  s,
  random = FALSE,
  Nrep = 1,
  Nexact = 1e+06
)

representative_poipower(
  n,
  trunc = 0,
  s,
  eta,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000,
  constant = "G25_K51"
)

representative_ls(
  n,
  trunc = NULL,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  alpha,
  N,
  s,
  oM,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000
)

representative_powbend(
  n,
  trunc = NULL,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  omega,
  oM,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000
)

representative_poilog(
  n,
  trunc = 0,
  mu = 0,
  sig = 1,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000,
  constant = "G25_K51"
)

representative_nblog(
  n,
  trunc = 0,
  mu = 0,
  sig = 1,
  odisp = 1,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000,
  maxit = 16,
  constant = "G25_K51"
)

representative_poipowbend(
  n,
  trunc = 0,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  eta = 0.001,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000,
  exact = 1e+05,
  constant = "G25_K51"
)

representative_nbpowbend(
  n,
  trunc = 0,
  s,
  lambda = 0.01,
  minlogLambda = -log(lambda),
  eta = 0.001,
  odisp = 1,
  random = FALSE,
  Nrep = 1,
  Nexact = 10000,
  maxit = 16,
  exact = 1e+05,
  constant = "G25_K51"
)

representative_poils(...)

representative_nbls(...)
}
\arguments{
\item{n}{the number of observed species to sample from SAD}

\item{trunc}{the truncation point only above which the species are observed}

\item{s}{the exponent of the denominator x^s}

\item{random}{logical; when \code{TRUE}, the first RAD is randomly (instead of evenly) sampled from the SAD}

\item{Nrep}{the number of replicates to be generated}

\item{Nexact}{the maximum abundance beyond which the RAD is approximated}

\item{eta}{the coefficient of sampling effort as the mean of the compounding Poisson/negative-binomial distribution}

\item{constant}{a character name for the Gauss-Kronrod system or a list of user-supplied numeric vectors; see \code{integral_refined()} for more information}

\item{lambda, minlogLambda}{the coefficient of the exponent of the numerator e^(-lambda*x)}

\item{alpha, N}{the alternative parameterization of log-series that is compatible with \code{sads::fitls}}

\item{oM}{alternative parameterization of \code{minlogLambda} that is compatible with \code{sads::fitpowbend}}

\item{odisp}{the coefficient of sampling bias as the overdispersion in negative-binomial distribution}

\item{maxit}{the maximum iteration before bounded quantile search is terminated}

\item{exact}{the number of terms to be summed in the infinite series before approximation applies}
}
\value{
\verb{representative_*} functions return a list of RAD whose elements are sorted numeric values
}
\description{
A set of functions for fast RAD approximation from SAD models
}
\details{
\verb{representative_*} are a set of functions dedicated to generating RAD replicates from SAD models to evaluate the absolute goodness of fit

They apply a hybrid strategy of exact and approximate quantile to balance the accuracy and time complexity of prediction. Specifically, larger quantiles are approximated by a simpler distribution for faster prediction without losing too much accuracy.
}
