% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octave_fit_test.R
\name{sad2octave}
\alias{sad2octave}
\title{Convert SAD fit to octave fit}
\usage{
sad2octave(x, model, params, pred_rad, fixed = 0)
}
\arguments{
\item{x}{the species abundance for each species}

\item{model}{the base name of model (e.g., poipowbend for Poisson-powerbend)}

\item{params}{the fitted full parameters of the model (e.g., from the \code{fullcoef} slot of a \code{mle2} class object)}

\item{pred_rad}{the predicted species abundance using the fitted model (e.g., from representative_* functions)}

\item{fixed}{the number of fixed parameters for calculating AIC and BIC}
}
\value{
\code{sad2octave} returns a list containing the SAD in octave representation and the corresponding fitting metrics if a SAD fit is provided
}
\description{
Convert the SAD fit to octave fit to allow comparison with octave-based models like gambin
}
\details{
The octaves start with '0', which contains all the species with 1 individuals. Octave 'i' contains species with \link{2^i,2^(i+1)-1} individuals.

For SAD in octaves, the truncation is fixed to be 0.
}
